Restructured the project to separate out Commands and Dr. Egeria (md_processing) from core pyegeria.
