+++ commands/cli/egeria.py
@@ -4,7 +4,7 @@ SPDX-License-Identifier: Apache-2.0
 Copyright Contributors to the ODPi Egeria project.
 
 
-A command line interface for Egeria Users - all commands
+A command line interface for Egeria Users - all md_commands
 
 This is an emerging capability based on the **click** package. Feedback welcome!
 
@@ -1032,7 +1032,7 @@ def show_cat(ctx):
 @show_cat.group("info")
 @click.pass_context
 def show_cat_info(ctx):
-    """Group of commands to show information about various Egeria objects"""
+    """Group of md_commands to show information about various Egeria objects"""
     pass
 
 
@@ -1109,7 +1109,7 @@ def show_asset_graph(ctx, asset_guid):
 
     Usage: show asset-graph <asset-guid>
 
-           asset-guid must be a valid asset guid. These can be found through other commands such as 'show tech-type-elements'
+           asset-guid must be a valid asset guid. These can be found through other md_commands such as 'show tech-type-elements'
 
     """
     c = ctx.obj
@@ -1583,7 +1583,7 @@ def show_asset_graph(ctx, asset_guid):
 
     Usage: show asset-graph <asset-guid>
 
-           asset-guid must be a valid asset guid. These can be found through other commands such as 'show tech-type-elements'
+           asset-guid must be a valid asset guid. These can be found through other md_commands such as 'show tech-type-elements'
 
     """
     c = ctx.obj
@@ -1781,7 +1781,7 @@ def show_ops(ctx):
 @show_ops.group("repository")
 @click.pass_context
 def show_repo(ctx):
-    """Group of commands to show repository information"""
+    """Group of md_commands to show repository information"""
     pass
 
 
@@ -1803,7 +1803,7 @@ def show_archives(ctx):
 @show_ops.group("platforms")
 @click.pass_context
 def show_platform(ctx):
-    """Group of commands to show information about Egeria platforms"""
+    """Group of md_commands to show information about Egeria platforms"""
     pass
 
 
@@ -1820,7 +1820,7 @@ def show_platform_status(ctx):
 @show_ops.group("servers")
 @click.pass_context
 def show_server(ctx):
-    """Group of commands to show information about Egeria servers"""
+    """Group of md_commands to show information about Egeria servers"""
     pass
 
 
@@ -1864,7 +1864,7 @@ def show_startup_status(ctx):
 @show_ops.group("engines")
 @click.pass_context
 def engine_host(ctx):
-    """Group of commands to show information about Egeria engines"""
+    """Group of md_commands to show information about Egeria engines"""
     pass
 
 
@@ -1943,7 +1943,7 @@ def eng_activity_status(ctx, rowlimit: int, list: bool, compressed: bool):
 @show_ops.group("integrations")
 @click.pass_context
 def integrations(ctx):
-    """Group of commands to show information about Egeria integrations"""
+    """Group of md_commands to show information about Egeria integrations"""
     pass
 
 
@@ -2010,7 +2010,7 @@ def tell_ops(ctx):
 @tell_ops.group("integration-daemon")
 @click.pass_context
 def tell_integration_daemon(ctx):
-    """Group of commands to an integration_daemon"""
+    """Group of md_commands to an integration_daemon"""
     pass
 
 
@@ -2066,7 +2066,7 @@ servers.add_command(stop_server)
 @tell_ops.group("engine-host")
 @click.pass_context
 def engine_host(ctx):
-    """Group of commands to an engine-host"""
+    """Group of md_commands to an engine-host"""
     pass
 
 
@@ -2078,7 +2078,7 @@ engine_host.add_command(refresh_gov_eng_config)
 @tell_ops.group("repository")
 @click.pass_context
 def repository(ctx):
-    """Group of commands to a repository"""
+    """Group of md_commands to a repository"""
     pass
