+++ md_processing/md_commands/blueprint_commands.py
@@ -6,19 +6,16 @@ from typing import Optional
 from rich.markdown import Markdown
 
 from pyegeria.egeria_tech_client import EgeriaTech
-from md_processing.utils.common_utils import (
-    debug_level, print_msg, ALWAYS, ERROR, INFO, pre_command, EXISTS_REQUIRED
+from md_processing.md_processing_utils.common_utils import (
+    debug_level, print_msg, update_element_dictionary
 )
-from md_processing.utils.extraction_utils import (
-    extract_command_plus, process_simple_attribute, process_name_list
+from md_processing.md_processing_utils.extraction_utils import (
+    extract_command_plus, process_simple_attribute, process_name_list, process_element_identifiers, update_a_command
 )
-from md_processing.utils.validation_utils import (
-    process_element_identifiers, update_a_command
-)
-from md_processing.utils.display_utils import (
-    BLUEPRINT_NAME_LABELS, COMPONENT_NAME_LABELS
+
+from md_processing.md_processing_utils.md_processing_constants import (
+    BLUEPRINT_NAME_LABELS, COMPONENT_NAME_LABELS, ALWAYS, ERROR, INFO, pre_command, EXISTS_REQUIRED,
     )
-from pyegeria.dr_egeria_state import update_element_dictionary
 
 def process_blueprint_upsert_command(egeria_client: EgeriaTech, txt: str, directive: str = "display") -> Optional[str]:
     """
@@ -30,7 +27,7 @@ def process_blueprint_upsert_command(egeria_client: EgeriaTech, txt: str, direct
     :param directive: an optional string indicating the directive to be used - display, validate or execute
     :return: A string summarizing the outcome of the processing.
     """
-    from md_processing.utils.common_utils import set_debug_level
+    from md_processing.md_processing_utils.common_utils import set_debug_level
     
     command, object_type, object_action = extract_command_plus(txt)
     set_debug_level(directive)
@@ -146,7 +143,7 @@ def process_solution_component_upsert_command(egeria_client: EgeriaTech, txt: st
     :param directive: an optional string indicating the directive to be used - display, validate or execute
     :return: A string summarizing the outcome of the processing.
     """
-    from md_processing.utils.common_utils import set_debug_level
+    from md_processing.md_processing_utils.common_utils import set_debug_level
     
     command, object_type, object_action = extract_command_plus(txt)
     set_debug_level(directive)
